requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
SQLAlchemy>=2.0
pymysql>=1.1
pydantic-settings>=2.0
python-jose[cryptography]>=3.3
passlib>=1.7
python-multipart>=0.0.9
jinja2>=3.1
